initial fix/refactor of private credentials
