+++ .gitmodules
@@ -1,7 +1,6 @@
 [submodule "software-communication"]
 	path = software-communication
 	url = ../software-communication.git
-	# url = git@github.com:SSL-A-Team/software-communication.git
-[submodule "robot-private-data"]
-	path = robot-private-data
-	url = git@github.com:SSL-A-Team/robot-private-data.git
+[submodule "ateam-credentials/src/private_credentials"]
+	path = ateam-credentials/src/private_credentials
+	url = git@github.com:SSL-A-Team/ateam-private-credentials.git
