+++ ateam-credentials/.gitignore
