+++ ateam-credentials/src/lib.rs
@@ -1,0 +1,31 @@
+#![no_std]
+#![allow(non_upper_case_globals)]
+
+pub mod public_credentials;
+#[cfg(not(feature = "no-private-credentials"))]
+pub mod private_credentials;
+
+#[derive(Clone, Copy)]
+pub enum WifiCredentialIndices {
+    DummyField = 0,
+    DummyDevelopmentNetwork = 1,
+    COUNT = 2,
+}
+
+#[derive(Clone, Copy)]
+pub struct WifiCredential {
+    ssid: &'static str,
+    password: &'static str,
+}
+
+impl WifiCredential {
+    pub fn get_ssid(&self) -> &'static str {
+        self.ssid
+    }
+
+    pub fn get_password(&self) -> &'static str {
+        self.password
+    }
+}
+
+
