+++ ateam-credentials/src/public_credentials/mod.rs
@@ -1,0 +1,1 @@
+pub mod wifi;
