+++ ateam-credentials/src/public_credentials/wifi.rs
@@ -1,27 +1,10 @@
-#![no_std]
-#![allow(non_upper_case_globals)]
+use WifiCredential;
 
 #[derive(Clone, Copy)]
 pub enum WifiCredentialIndices {
     DummyField = 0,
     DummyDevelopmentNetwork = 1,
     COUNT = 2,
-}
-
-#[derive(Clone, Copy)]
-pub struct WifiCredential {
-    ssid: &'static str,
-    password: &'static str,
-}
-
-impl WifiCredential {
-    pub fn get_ssid(&self) -> &'static str {
-        self.ssid
-    }
-
-    pub fn get_password(&self) -> &'static str {
-        self.password
-    }
 }
 
 pub static wifi_credentials: [WifiCredential; WifiCredentialIndices::COUNT as usize] = [
